--- blocks/breadcrumb/breadcrumb.css
@@ -1,31 +1,0 @@
-main > .breadcrumb-container {
-    padding: 10px;
-    grid-row: 1;
-    grid-column: 1;
-  }
-  
-  .blog main div.breadcrumb-container {
-    padding: 10px;
-  }
-  
-  @media (min-width: 900px) {
-    main .breadcrumb-container {
-      grid-row: 1;
-      grid-column: 1 / span 2;
-    }
-  }
-  
-  .breadcrumb > nav {
-    color: var(--link-color);
-    font-size: var(--body-font-size-xs);
-    padding-left: 10px;
-    min-height: 26px;
-  }
-  
-  .breadcrumb > nav a {
-    text-decoration: none;
-  }
-  
-  .breadcrumb .breadcrumb-separator {
-    margin-inline: 1.5ch;
-  }
