--- blocks/embed/embed.css
@@ -1,65 +1,0 @@
-main .embed {
-    width: unset;
-    text-align: center;
-    max-width: 800px;
-    margin: 32px auto;
-  }
-  
-  main .embed > div {
-    display: flex;
-    justify-content: center;
-  }
-  
-  main .embed.embed-twitter .twitter-tweet-rendered {
-    margin-left: auto;
-    margin-right: auto;
-  }
-  
-  main .embed .embed-placeholder {
-    width: 100%;
-    aspect-ratio: 16 / 9;
-    position: relative;
-  }
-  
-  main .embed .embed-placeholder > * {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-  }
-  
-  main .embed .embed-placeholder picture img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-  
-  main .embed .embed-placeholder-play button {
-    box-sizing: border-box;
-    position: relative;
-    display: block;
-    transform: scale(3);
-    width: 22px;
-    height: 22px;
-    border: 2px solid;
-    border-radius: 20px;
-    padding: 0;
-  }
-  
-  main .embed .embed-placeholder-play button::before {
-    content: "";
-    display: block;
-    box-sizing: border-box;
-    position: absolute;
-    width: 0;
-    height: 10px;
-    border-top: 5px solid transparent;
-    border-bottom: 5px solid transparent;
-    border-left: 6px solid;
-    top: 4px;
-    left: 7px;
-  }
