--- blocks/fragment/fragment.css
@@ -1,1 +1,0 @@
-
