+++ styles/styles.css
@@ -14,10 +14,9 @@
   /* colors */
   --link-color: #035fe6;
   --link-hover-color: #136ff6;
-  --background-color: #ffffff;
+  --background-color: #fff;
   --overlay-background-color: #eee;
   --highlight-background-color: #ccc;
-  --green-background-color:#0fc7d4;
   --text-color: #000;
 
   /* fonts */
@@ -224,7 +223,7 @@ main img {
   width: 24px;
 }
 
-.icon svg {
+.icon img {
   height: 100%;
   width: 100%;
 }
@@ -249,7 +248,4 @@ main .section {
 /* section metadata */
 main .section.highlight {
   background-color: var(--highlight-background-color);
-}
-main .section.green {
-  background-color: var(--green-background-color);
 }
