--- tools/sidekick/config.json
@@ -1,14 +1,0 @@
-{
-    "project": "My First Franklin",
-    "plugins": [
-      {
-        "id": "asset-library",
-        "title": "AEM Assets Library",
-        "environments": ["edit"],
-        "url": "https://experience.adobe.com/solutions/CQ-assets-selectors/static-assets/resources/franklin/asset-selector.html",
-        "isPalette": true,
-        "includePaths": ["**.docx**"],
-        "paletteRect": "top: 50px; bottom: 10px; right: 10px; left: auto; width:400px; height: calc(100vh - 60px)"
-      }
-    ]
-  }
